backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# PDF processing
PyMuPDF==1.23.14
pillow==10.1.0

# OCR engines
paddleocr==2.7.3
# doctr[torch]==0.7.0  # Alternative OCR, uncomment if needed

# Document processing
unstructured[pdf]==0.11.8
numpy==1.25.2

# ML/NLP for semantic matching
sentence-transformers==2.2.2
torch==2.1.2
rapidfuzz==3.6.1

# Optional cloud providers (install only if needed)
# google-cloud-documentai==2.20.1
# boto3==1.34.34
# azure-ai-formrecognizer==3.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
coverage==7.3.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1